Renamed files for clean code
