+++ src/index.js
@@ -1,6 +1,6 @@
 import './style.css';
 
-import getMovies from './modules/component.js';
+import getMovies from './modules/renderMovies.js';
 
 window.addEventListener('DOMContentLoaded', () => {
   // listContainer.innerHTML = ' ';
